new sportsman

+++ contracts/MedalVerse.sol
@@ -19,13 +19,12 @@ contract MedalVerse is
 	OrganizerHandler,
 	MedalHandler
 {
-
 	IERC20 internal Token;
-	
-	constructor(address addressToken)  {
+
+	constructor(address addressToken) {
 		Token = IERC20(addressToken);
 	}
-		
+
 	// Modifiers ----------------------------
 	modifier isNotNull(address a)
 		virtual
@@ -72,7 +71,7 @@ contract MedalVerse is
 		string memory _iconURI,
 		string memory _userName,
 		string memory _email,
-		uint8 _role,
+		uint256 _role,
 		uint256 _sportsCategory
 	) public onlyOwner isNotNull(_userAddress) {
 		addUser(_userAddress, _iconURI, _userName, _email, _role);
@@ -201,12 +200,9 @@ contract MedalVerse is
 		eventSetPosition(eventID, posX, posY);
 	}
 
-	///@dev Withdraw token in the contract
+	///@dev Withdraw token placed in the contract
 	function withdraw() public onlyOwner {
-		uint balance = Token.balanceOf(address(this)); // check balance of contract MedalVerse
-		Token.transfer(msg.sender, balance); // transfer balance to owner 
+		uint256 balance = Token.balanceOf(address(this)); // check balance of contract MedalVerse
+		Token.transfer(msg.sender, balance); // transfer balance to owner
 	}
-	
 }
-
-
